requirements.txt
fastapi
uvicorn
google-genai==1.67.0
python-dotenv
pydantic